+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## [5.1.1] - --
+## [6.0.0] - --
 
+- Changed `ManagedPool.Shared` to a new class `SharedPool`.
+- Changed `ManagedPoolBase` to `ManagedPool`, alongside any API breaking changes.
 - Fixed unloading of pools that are not fully loaded being inefficient.
 
 ## [5.1.0] - 2022-04-03